Away!
A whole lotta shit.
  -Redid the way a tab is laid out; there is now a hidden text field for the away msg
  -Detects /away (reason sets away, no reason returns)
  -Toggles the toolbar image & tooltip text

Most importantly (and frustraingly) does this for the window of a server
and ALL of its channel tabs.

On a side note, it added about 7 new functions some of which may not be needed
or can be merged. At very least they need to be sorted within the file