Working on NickList modes
The old way (just using nickList with special characters), if a user
had '+', got '%', then lost '%', there would be no way to know he
still had voice.

Now it is kept inside a "SpecialUsers" array of lists.

When a user changes nicks or modes, it just modifies the lists and
searches through for the highest mode, then adds that prefix to nickList.